corse_서비스/Service/flask/requirements.txt
click==7.1.2
cycler==0.11.0
et-xmlfile==1.1.0
Flask==1.1.2
fonttools==4.28.2
greenlet==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.2
mariadb==1.0.8
MarkupSafe==1.1.1
matplotlib==3.5.0
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
PyMySQL==1.0.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
uWSGI==2.0.19.1
Werkzeug==1.0.1